Deduplicate the copy pasted algorithm classes, fix all fixable warnings
In each of the one_to_one, one_to_many, n_to_mn and n_to_m groups,
remove the duplicated code from all the naive algorithms and create a
intermeddiate predecesor class naive_gpu_x_to_x class. This class
defines the CPU and GPU buffers and implements allocation, data loading,
trasnfer to and from GPU and deallocation.

Change labels from vector of strings into static C array of const char*.
This way there is no initialization or allocation that could throw
during applciation startup and which we could not catch, so clang stops
warning us about that.
The static array size can be retrieved using std::size(), and vector
constructed from it using std::begin() adn std::end().

Changed labels variable to inline definition instead of out of line
definition. This makes it more readable and saves some lines.

Add nodiscard everywhere Clion told us to. Basically all getters etc.

Removed unused headers from all files.

Fix infinite recursion in naive_shuffle.cu, where the recursive template
call did not subtract from the template argument.

+++ src/naive_warp_per_shift_shared_mem.cu
@@ -1,8 +1,6 @@
-#include <cuda.h>
 #include <cuda_runtime.h>
 
 #include <cooperative_groups.h>
-#include <cooperative_groups/memcpy_async.h>
 #include <cooperative_groups/reduce.h>
 
 #include <cassert>
@@ -11,8 +9,6 @@
 #include "types.cuh"
 #include "cuda_helpers.cuh"
 #include "shared_mem.cuh"
-#include "row_distribution.cuh"
-#include "argument_error.hpp"
 #include "warp_size.hpp"
 
 namespace cg = cooperative_groups;
@@ -76,9 +72,9 @@ __device__ void compute_from_shared_mem_buffers(
     int buffer_offset = buffer_row_shift * static_cast<int>(row_size);
 
     for (
-        int right_idx = warp_right_start_offset + warp.thread_rank();
+        int right_idx = warp_right_start_offset + static_cast<int>(warp.thread_rank());
         right_idx < warp_right_end_offset;
-        right_idx += warp.size()
+        right_idx += static_cast<int>(warp.size())
     ) {
         auto l = left_buffer[right_idx + buffer_offset];
         for (dsize_t i = 0; i < NUM_RIGHT; ++i) {
